fix: chat restructure
